+++ R/report.doc.R
@@ -11,7 +11,7 @@
 .output=new.env(parent = emptyenv())
 
 
-#' Export a statistical table into 'Microsoft Word'
+#' Export a statistical table into a 'Microsoft Word' or a R markdown document
 #'
 #' @param table A desc object that report statistics (the results of \code{report.quanti} or \code{report.quali})
 #' @param title Character. The title of the table
@@ -29,12 +29,36 @@
 #' @description
 #' \code{report.doc} 
 #' This function enables to export the table created with \code{\link{report.quali}} \code{\link{report.quanti}} or \code{\link{report.lsmeans}}
-#' to a Microsoft Word in a "clinical standard" format. 
+#' to a Microsoft Word or a R markdown document. 
 #' 
-#' It's also possible to use it to have a preview of the table in HTML format if the doc argument is NULL.
+#' It's also possible to use report.doc to have a preview of the table in HTML format if the \code{doc} argument is NULL.
 #' 
 #' @details
-#' It creates a flextable object from a desc object and can eventually add it directly into a rdocx object.
+#' 
+#' This function creates a flextable object from a desc object.
+#' 
+#' \strong{For Microsoft Word documents:}
+#' 
+#' The argument \code{doc} should be used so the flextable is added to a rdocx object.
+#' 
+#' Like:
+#' 
+#' \code{doc=read_docx()}
+#' 
+#' \code{tab=report.quanti(data=data,y="y_numeric",x1="GROUP")}
+#' 
+#' \code{doc=report.doc(tab,doc=doc)}
+#' 
+#' \strong{For R markdown documents:}
+#' 
+#' Just don't use the \code{doc} argument. Something like:
+#' 
+#' ```\{r, include=TRUE\} \cr
+#' \code{tab=report.quanti(data=data,y="y_numeric",x1="GROUP")} \cr
+#' \code{doc=report.doc(tab)} \cr
+#' \code{doc} \cr
+#' ``` \cr
+#'  
 #' 
 #' @return  
 #' A flextable object (if doc=NULL) or a rdocx object (if doc= an rdocx object).
@@ -51,21 +75,22 @@
 #'library(lme4)
 #'library(nlme)
 #'
-#'data(data)
+#'data(datafake)
 #'
-#'tab=report.quanti(data=data,y="y_numeric",
+#'tab=report.quanti(data=datafake,y="y_numeric",
 #'		x1="GROUP",x2="TIMEPOINT",at.row="TIMEPOINT",subjid="SUBJID")
 #' 
 #'mod=glm(y_logistic~GROUP+TIMEPOINT+GROUP*TIMEPOINT,
-#'family=binomial,data=data,na.action=na.omit)
+#'family=binomial,data=datafake,na.action=na.omit)
+#' 
 #'test=emmeans(mod,~GROUP|TIMEPOINT)
-#'tab.mod=report.lsmeans(lsm=test,x1="GROUP",
-#'x2="TIMEPOINT",at.row="TIMEPOINT",data=data)
+#' 
+#'tab.mod=report.lsmeans(lsm=test)
 #' 
 #' 
 #'doc=read_docx()
 #'
-#'doc=body_add_par(doc,"A beautiful reporting using ClinReport", style = "heading 1")
+#'doc=body_add_par(doc,"A statistical report using ClinReport", style = "heading 1")
 #'
 #'doc=report.doc(tab,title="Quantitative statistics",
 #'		colspan.value="Treatment group",doc=doc,init.numbering=TRUE)
@@ -97,12 +122,12 @@
 #' # Load data
 #' #####################
 #' 
-#'data(data) 
-#'head(data)
+#'data(datafake) 
+#'head(datafake)
 #'
 #' # Removing baseline data for the model
 #'
-#'data.mod=droplevels(data[data$TIMEPOINT!="D0",])
+#'data.mod=droplevels(datafake[datafake$TIMEPOINT!="D0",])
 #' 
 #' #####################
 #' # Create your stats tables and graphics
@@ -112,7 +137,7 @@
 #' # since it's a big enough table, we don't want it to overlap 2 pasges
 #' # so we split it in two with split.desc function
 #'
-#'tab1=report.quanti(data=data,y="y_numeric",
+#'tab1=report.quanti(data=datafake,y="y_numeric",
 #'		x1="GROUP",x2="TIMEPOINT",at.row="TIMEPOINT",subjid="SUBJID")
 #' 
 #'
@@ -127,30 +152,30 @@
 #'
 #' # Qualitative stats (2 explicative variables) ##################################
 #' 
-#'tab2=report.quali(data=data,y="y_logistic",
-#'		x1="GROUP",x2="TIMEPOINT",at.row="TIMEPOINT",total=T,subjid="SUBJID")
+#'tab2=report.quali(data=datafake,y="y_logistic",
+#'		x1="GROUP",x2="TIMEPOINT",at.row="TIMEPOINT",total=TRUE,subjid="SUBJID")
 #'
 #' gg2=plot(tab2,title="Response distribution (%) by day and treatment group",
 #' legend.label="Y levels")
 #'
 #' # Qualitative stats (no explicative variable)  ###################################
 #' 
-#'tab3=report.quali(data=data,y="VAR",y.label="Whatever")
+#'tab3=report.quali(data=datafake,y="VAR",y.label="Whatever")
 #'
 #' # Qualitative stats (no explicative variables ; add number of subjects in header)#
 #' 
-#'tab4=report.quali(data=data,y="VAR",y.label="Whatever",
+#'tab4=report.quali(data=datafake,y="VAR",y.label="Whatever",
 #'		subjid="SUBJID")
 #'
 #' # Qualitative stats (1 explicative variable)#######################################
 #' 
-#'tab5=report.quali(data=data,y="VAR",y.label="Whatever",x1="GROUP",
+#'tab5=report.quali(data=datafake,y="VAR",y.label="Whatever",x1="GROUP",
 #'		subjid="SUBJID")
 #'
 #'
 #'# Quantitative stats (1 explicative variable)#######################################
 #'
-#'tab6=report.quanti(data=data,y="y_numeric",y.label="Whatever 2",x1="GROUP",
+#'tab6=report.quanti(data=datafake,y="y_numeric",y.label="Whatever 2",x1="GROUP",
 #'		subjid="SUBJID")
 #'
 #'# Quali-Quanti table
@@ -165,12 +190,11 @@
 #' 
 #'anov1=Anova(mod1)
 #'
-#'tab.mod1=report.lsmeans(lsm=test1,x1="GROUP",
-#' x2="TIMEPOINT",at.row="TIMEPOINT",data=data.mod)
+#'tab.mod1=report.lsmeans(lsm=test1)
 #'
 #'gg.mod1=plot(tab.mod1,title="LS-Means response evolution as a function of time\n
 #' by treatment group (95% CI)",
-#' legend.label="Treatment groups",ylab="Y mean",add.ci=T)
+#' legend.label="Treatment groups",ylab="Y mean",add.ci=TRUE)
 #' 
 #' # Linear model (1 group only): Anova and LS-Means and graph reporting ################
 #' 
@@ -179,11 +203,11 @@
 #'anov2=Anova(mod2,type=3)
 #' 
 #'test2=emmeans(mod2,~GROUP)
-#'tab.mod2=report.lsmeans(lsm=test2,x1="GROUP",data=data.mod)
+#'tab.mod2=report.lsmeans(lsm=test2)
 #'
 #'
 #'gg.mod2=plot(tab.mod2,title="LS-Means response\nby treatment group (95% CI)",
-#'		legend.label="Treatment groups",ylab="Y mean",add.ci=T)
+#'		legend.label="Treatment groups",ylab="Y mean",add.ci=TRUE)
 #'
 #' # Linear mixed model (order 2 interaction):
 #' # Anova and LS-Means and graph reporting #################
@@ -195,27 +219,27 @@
 #' 
 #'test3=emmeans(mod3,~GROUP|TIMEPOINT)
 #' 
-#'tab.mod3=report.lsmeans(lsm=test3,x1="GROUP",
-#' x2="TIMEPOINT",at.row="TIMEPOINT",data=data.mod)
+#'tab.mod3=report.lsmeans(lsm=test3)
 #'
 #'gg.mod3=plot(tab.mod3,title="LS-Means response evolution as a function of time\n
 #'by treatment group (95% CI Mixed model)",
-#'		legend.label="Treatment groups",ylab="Y mean",add.ci=T)
+#'		legend.label="Treatment groups",ylab="Y mean",add.ci=TRUE)
 #'
 #' # Contrast example
 #'
 #'contr=contrast(test3, "trt.vs.ctrl", ref = "A")
 #'
-#'tab.mod3.contr=report.lsmeans(lsm=contr,x1="TIMEPOINT",
-#'		data=data.mod,contrast=TRUE,at.row="contrast")
+#'tab.mod3.contr=report.lsmeans(lsm=contr)
 #'
 #'gg.mod3.contr=plot(tab.mod3.contr,title="LS-Means contrast versus reference A\n
 #'				(95% CI Mixed model)",
-#'		legend.label="Treatment groups",ylab="Y mean",add.ci=T,add.line=F)
+#'		legend.label="Treatment groups",ylab="Y mean",add.ci=TRUE,add.line=FALSE)
 #'
 #'
-#'
+#' ############################################################
 #' # Generalized Logistic Linear model (order 2 interaction):
+#' ############################################################
+#' 
 #' # Anova LS-Means and graph reporting ##########
 #' 
 #'mod4=glm(y_logistic~baseline+GROUP+TIMEPOINT+GROUP*TIMEPOINT,
@@ -225,12 +249,11 @@
 #' 
 #'test4=emmeans(mod4,~GROUP|TIMEPOINT)
 #'
-#'tab.mod4=report.lsmeans(lsm=test4,x1="GROUP",
-#' x2="TIMEPOINT",at.row="TIMEPOINT",data=data.mod)
+#'tab.mod4=report.lsmeans(lsm=test4,at.row="TIMEPOINT")
 #'
 #'gg.mod4=plot(tab.mod4,title="LS-Means response evolution as a function of time\n
 #'by treatment group (95% CI Logistic model)",
-#'		legend.label="Treatment groups",ylab="Y mean",add.ci=T)
+#'		legend.label="Treatment groups",ylab="Y mean",add.ci=TRUE)
 #'
 #' # Generalized Poisson Linear model (order 2 interaction):
 #' # Anova LS-Means and graph reporting #'
@@ -244,13 +267,12 @@
 #' 
 #'test5=emmeans(mod5,~GROUP|TIMEPOINT)
 #'
-#'tab.mod5=report.lsmeans(lsm=test5,x1="GROUP",
-#' x2="TIMEPOINT",at.row="TIMEPOINT",type="response",data=data.mod)
+#'tab.mod5=report.lsmeans(lsm=test5,at.row="TIMEPOINT")
 #'
 #'
 #'gg.mod5=plot(tab.mod5,title="LS-Means response evolution as a function of time\n
 #'by treatment group (95% CI Poisson model)",
-#'		legend.label="Treatment groups",ylab="Y mean",add.ci=T)
+#'		legend.label="Treatment groups",ylab="Y mean",add.ci=TRUE)
 #' 
 #' #####################
 #' # Create your report
@@ -266,8 +288,8 @@
 #'doc=body_add_par(doc,"Descriptive statistics", style = "heading 2")
 #'
 #'doc=report.doc(tab1.1,title="Quantitative statistics (2 explicative variables) (Table 1/2)",
-#'		colspan.value="Treatment group",doc=doc,init.numbering=T,
-#' page.break=F)
+#'		colspan.value="Treatment group",doc=doc,init.numbering=TRUE,
+#' page.break=FALSE)
 #'
 #'doc=report.doc(tab1.2,title="Quantitative statistics (2 explicative variables) (Table 2/2)",
 #'		colspan.value="Treatment group",doc=doc)
@@ -396,8 +418,8 @@ report.doc <- function(table,...)
 #' @rdname report.doc
 #' @export 
 
-report.doc.desc=function(table,title,colspan.value=NULL,doc=NULL,
-		init.numbering=F,numbering=T,font.name="Times",page.break=T,font.size=11,...)
+report.doc.desc=function(table,title=NULL,colspan.value=NULL,doc=NULL,
+		init.numbering=F,numbering=T,font.name="Times",page.break=T,font.size=10,...)
 {
 	
 	
@@ -434,7 +456,14 @@ report.doc.desc=function(table,title,colspan.value=NULL,doc=NULL,
 	
 	if(numbering)
 	{
-		title= paste0("Output ",get("number",envir=.output),": ",c(title))
+		if(!is.null(title))
+		{
+			title= paste0("Table ",get("number",envir=.output),": ",c(title))
+		}else
+		{
+			title= paste0("Table ",get("number",envir=.output),": ",c(table$title))
+		}
+		
 	}
 	
 	
@@ -571,6 +600,7 @@ report.doc.desc=function(table,title,colspan.value=NULL,doc=NULL,
 }
 
 #' @param type.anova Passed to \code{Anova} function from car package (see its documentation).
+#' @param pretty.label Logical. Default to FALSE. If TRUE, use the function \code{make.label} with default option on the rownames of the anova table 
 #' 
 #' @importFrom xtable xtable
 #' @rdname report.doc
@@ -579,7 +609,7 @@ report.doc.desc=function(table,title,colspan.value=NULL,doc=NULL,
 
 
 report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,numbering=T,
-		init.numbering=F,font.name="Times",font.size=11,page.break=T,...)
+		init.numbering=F,font.name="Times",font.size=10,page.break=T,pretty.label=FALSE,...)
 {
 	
 	
@@ -596,7 +626,7 @@ report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,number
 	
 	if(numbering)
 	{
-		title= paste0("Output ",get("number",envir=.output),": ",c(title))
+		title= paste0("Table ",get("number",envir=.output),": ",c(title))
 	}
 	
 	
@@ -606,6 +636,12 @@ report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,number
 	
 	
 	ncol=ncol(as.data.frame(table))
+	
+	
+	if(pretty.label)
+	{
+		rownames(table)=make.label(rownames(table))
+	}
 	
 	xtab=xtable(table)
 	ft=xtable_to_flextable(xtab,NA.string = "-")
@@ -631,8 +667,8 @@ report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,number
 	
 	# change font 
 	
-	ft=font(ft,fontname=font.name,part ="all")
-	ft=fontsize(ft,size=font.size,part ="all")
+	ft=flextable::font(ft,fontname=font.name,part ="all")
+	ft=flextable::fontsize(ft,size=font.size,part ="all")
 	
 	if(!is.null(doc))
 	{